Revert "chore: Deploy changes for focus point (#65)"
This reverts commit 19521ec917896f128f0ed2f63a59c77fb36bbc71.
