+++ package.json
@@ -27,17 +27,16 @@
     "@types/lodash.sortby": "^4.7.6",
     "@types/logfmt": "^1.2.1",
     "@types/lz-string": "^1.3.34",
-    "@types/mapbox__polyline": "^1.0.2",
     "@types/node": "^14.0.24",
     "@types/node-fetch": "^2.5.7",
     "@types/uuid": "^8.0.0",
     "copyfiles": "^2.3.0",
     "jest": "^26.1.0",
     "nodemon": "^2.0.4",
-    "prettier": "^2.2.1",
-    "ts-jest": "^26.4.4",
+    "prettier": "^2.0.5",
+    "ts-jest": "^26.1.3",
     "ts-node": "^8.10.2",
-    "typescript": "^4.1.3"
+    "typescript": "^3.9.7"
   },
   "dependencies": {
     "@badrap/result": "^0.2.6",
@@ -48,7 +47,6 @@
     "@hapi/hapi": "^19.2.0",
     "@hapi/inert": "^6.0.1",
     "@hapi/vision": "^6.0.0",
-    "@mapbox/polyline": "^1.1.1",
     "@opencensus/core": "0.0.22",
     "@opencensus/exporter-stackdriver": "0.0.22",
     "@types/continuation-local-storage": "^3.2.2",
