+++ src/graphql/graphql-client.ts
@@ -1,4 +1,4 @@
-import { InMemoryCache, NormalizedCacheObject } from 'apollo-cache-inmemory';
+import { InMemoryCache } from 'apollo-cache-inmemory';
 import ApolloClient, { DefaultOptions } from 'apollo-client';
 import { HttpLink } from 'apollo-link-http';
 import fetch from 'node-fetch';
@@ -35,5 +35,3 @@ export default new ApolloClient({
   cache,
   defaultOptions
 });
-
-export type GraphQLClient = ApolloClient<NormalizedCacheObject>;
