+++ src/index.ts
@@ -22,8 +22,6 @@ import registerMetricsExporter from './utils/metrics';
 
 import { GaxiosError } from 'gaxios';
 import { PubSub } from '@google-cloud/pubsub';
-import serviceJourneyRoutes from './api/servicejourney';
-import serviceJourneyService from './service/impl/service-journey';
 
 process.on('unhandledRejection', err => {
   console.error(err);
@@ -58,7 +56,6 @@ process.on('unhandledRejection', err => {
     stopsRoutes(server)(stopsService(enturService, pubSubClient));
     geocoderRoutes(server)(geocoderService(enturService, pubSubClient));
     journeyRoutes(server)(js);
-    serviceJourneyRoutes(server)(serviceJourneyService(enturService));
     enrollmentRoutes(server)();
 
     registerMetricsExporter(projectId);
