--- src/service/impl/service-journey/service-journey-map.graphql
@@ -1,24 +1,0 @@
-query MapInfoByServiceJourneyId(
-  $serviceJourneyId: String!
-  $fromQuayId: String!
-  $toQuayId: String!
-) {
-  serviceJourney(id: $serviceJourneyId) {
-    pointsOnLink {
-      length
-      points
-    }
-    line {
-      transportMode
-      transportSubmode
-    }
-  }
-  fromQuay: quay(id: $fromQuayId) {
-    latitude
-    longitude
-  }
-  toQuay: quay(id: $toQuayId) {
-    latitude
-    longitude
-  }
-}
