+++ src/service/types.ts
@@ -5,14 +5,11 @@ import {
   Feature,
   StopPlaceDetails,
   Quay,
-  Departure,
-  LegMode,
-  PointsOnLink
+  Departure
 } from '@entur/sdk';
 import { FetchError } from 'node-fetch';
 import { boomify } from '@hapi/boom';
 import { CursoredQuery } from './cursored';
-import { TransportSubmode } from '../graphql/types';
 
 export interface Coordinates {
   latitude: number;
@@ -42,10 +39,6 @@ export type FeaturesQuery = {
   layers?: string[];
   tariff_zone_authorities?: string;
   limit?: number;
-
-  'focus.weight'?: number;
-  'focus.function'?: 'linear' | 'exp';
-  'focus.scale'?: 'string';
 };
 
 export interface QuaysForStopPlaceQuery {
@@ -177,11 +170,6 @@ export interface DeparturesForServiceJourneyQuery {
   date?: Date;
 }
 
-export interface ServiceJourneyMapInfoQuery {
-  fromQuayId?: string;
-  toQuayId?: string;
-}
-
 export interface DeparturesBetweenStopPlacesParams {
   limit?: number;
   start?: Date;
@@ -245,19 +233,6 @@ export type DepartureRealtimeData = {
 
 export type DeparturesRealtimeData = {
   [quayId: string]: DepartureRealtimeData;
-};
-
-export type MapLeg = {
-  mode?: LegMode;
-  faded?: boolean;
-  transportSubmode?: TransportSubmode;
-  pointsOnLink: PointsOnLink;
-};
-
-export type ServiceJourneyMapInfoData = {
-  mapLegs: MapLeg[];
-  start?: Coordinates;
-  stop?: Coordinates;
 };
 
 export class APIError extends Error {
