+++ swagger.yaml
@@ -340,28 +340,6 @@ paths:
           schema:
             type: string
           description: Successful
-  '/bff/v1/servicejourney/{id}/polyline':
-    get:
-      summary: Get departures for Service Journey
-      operationId: getBffV1ServicejourneyIdPolyline
-      parameters:
-        - type: string
-          name: id
-          in: path
-          required: true
-        - type: string
-          name: fromQuayId
-          in: query
-        - type: string
-          name: toQuayId
-          in: query
-      tags:
-        - bff
-      responses:
-        default:
-          schema:
-            type: string
-          description: Successful
   '/bff/v1/stop/{id}/departures':
     get:
       summary: >-
@@ -458,7 +436,7 @@ paths:
           in: query
         - type: string
           format: date
-          default: '2021-01-27T09:48:19.330Z'
+          default: '2020-12-07T13:56:39.320Z'
           name: startTime
           in: query
         - type: number
@@ -502,7 +480,7 @@ paths:
           in: query
         - type: string
           format: date
-          default: '2021-01-27T09:48:19.354Z'
+          default: '2020-12-07T13:56:39.338Z'
           name: startTime
           in: query
         - type: number
